refactor xdist logic away
